upgrade - fixed mistakes; added DTO's

+++ src/main/java/com/example/lista2ts/service/BookService.java
@@ -1,11 +1,15 @@
 package com.example.lista2ts.service;
 
+import com.example.lista2ts.dto.BookCreateDTO;
+import com.example.lista2ts.dto.BookCreateResponseDTO;
+import com.example.lista2ts.dto.BookDTO;
 import com.example.lista2ts.entity.BookEntity;
 import com.example.lista2ts.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class BookService {
@@ -16,19 +20,36 @@ public class BookService {
         this.bookRepository = bookRepository;
     }
 
-    public List<BookEntity> getAll() {
-        return bookRepository.findAll();
+    public List<BookDTO> getAll() {
+        var books = bookRepository.findAll();
+
+        return books.stream().map((book) -> new BookDTO(book.getId(), book.getIsbn(), book.getTitle(), book.getAuthor(), book.getPublisher(), book.getYearOfPublish(), book.getAvailableCopies())).collect(Collectors.toList());
     }
 
-    public BookEntity getOne(long bookId) {
-        return bookRepository.findById(bookId).orElseThrow(() -> new RuntimeException("Book not found!"));
+    public BookDTO getOne(long bookId) {
+        var book = bookRepository.findById(bookId).orElseThrow(() -> new RuntimeException("Book not found!"));
+        return new BookDTO(book.getId(), book.getIsbn(), book.getTitle(), book.getAuthor(), book.getPublisher(), book.getYearOfPublish(), book.getAvailableCopies());
     }
 
-    public BookEntity create(BookEntity book) {
-        return bookRepository.save(book);
+    public BookCreateResponseDTO create(BookCreateDTO book) {
+        var bookEntity = new BookEntity();
+
+        bookEntity.setAuthor(book.getAuthor());
+        bookEntity.setTitle(book.getTitle());
+        bookEntity.setIsbn(book.getIsbn());
+        bookEntity.setPublisher(book.getPublisher());
+        bookEntity.setAvailableCopies(book.getAvailableCopies());
+        bookEntity.setYearOfPublish(book.getYearOfPublish());
+
+        var newBook = bookRepository.save(bookEntity);
+
+        return new BookCreateResponseDTO(newBook.getId(), newBook.getIsbn(), newBook.getTitle(), newBook.getAuthor(), newBook.getPublisher(), newBook.getYearOfPublish(), newBook.getAvailableCopies());
     }
 
     public void delete(long bookId){
+        if (!bookRepository.existsById(bookId)){
+            throw new RuntimeException();
+        }
         bookRepository.deleteById(bookId);
     }
 }
